cli: use bufio directly, without textproto
This has practically the same code complexity but drops one indirect and
a dependency on the net package (here, other parts of the code still
depend on net via textproto).

I don't want the package to depend too much on cgo, which is often
required for the net package (DNS).