Drop generation number, only track inode
Getting the generation number turned out to be a bit more difficult: it
needs a real file descriptor (not one opened using O_PATH), and the test
was failing somehow when using something other than O_PATH.

Besides, in the new statx() proposal [1], there is a field for the birth
time (creation time), but not for version/generation numbers. So it
seems better to expect this syscall to land in Linux sometime so we can
use the inode+btime combination as unique identification. This works on
many other systems as well: many *BSD versions, darwin (OS X), and
Windows support creation times so it's better to standardise on that.
Generation numbers would be Linux-specific and always kind of a hack.

[1] https://lwn.net/Articles/707602/