Fix deadlock when scan job is marked finished before SCANOPTS is received
Finishing a scan job doesn't wait for SCANOPTS to be processed in a
goroutine. That means it is possible for the remote to receive SCANOPTS
and do the scan and send all status data. At this point, the client may
mark the scan job as finished, even though the SCANOPTS from the remote
hasn't even been sent to the client. Once the SCANOPTS is received, the
client doesn't know the current job. An error value is set, but
otherwise the system hangs.

This commit fixes the issue.